move files to single directory
