+++ src/components/pageWrap/PageWrap.jsx
@@ -2,32 +2,14 @@ import React from 'react';
 import Helmet from 'react-helmet';
 import PropTypes from 'prop-types';
 
-import { polyfillServiceUrl } from '../util/browserPolyfill';
+import { polyfillServiceUrl } from './browserPolyfill';
 
 /*
  * This just imports a blob of JS that New Relic give us so we can track usage/errors from real
  * users. We should only need to upgrade it in the event New Relic makes a major change, which
  * they'll let us know about.
  */
-import { newrelicBrowserJS } from '../util/newrelicBrowser';
-
-/*
- * -- Inline SVG icon sprite --
- *
- * raw SVG sprite from `swarm-icons`
- */
-const iconSpriteStyle = { display: 'none' };
-const iconSprite = require('raw-loader!swarm-icons/dist/sprite/sprite.inc');
-
-/*
- * Swarm logos
- */
-const swarmFavicon = require('file-loader!../assets/favicon.ico');
-const swarmIcon120x120 = require('file-loader!../assets/logos/m_swarm_120x120.png');
-const swarmIcon128x128 = require('file-loader!../assets/logos/m_swarm_128x128.png');
-const swarmIcon152x152 = require('file-loader!../assets/logos/m_swarm_152x152.png');
-const swarmIcon180x180 = require('file-loader!../assets/logos/m_swarm_180x180.png');
-const swarmIcon196x196 = require('file-loader!../assets/logos/m_swarm_196x196.png');
+import { newrelicBrowserJS } from './newrelicBrowser';
 
 /**
  * This component wraps all pages on the website, and through [Helmet](https://github.com/nfl/react-helmet/)
@@ -55,30 +37,14 @@ class PageWrap extends React.Component {
 	 * @return {React.element} the page wrapping component
 	 */
 	render() {
-		const { localeCode, baseCSSHref, webfontCSSHref } = this.props;
+		const { localeCode, baseCSSHref, webfontCSSHref, iconSprite } = this.props;
 
 		// Parse localeCode for ISO 639-1 languages code.
 		// (ie. 'en', 'it', etc)
 		// @see https://github.com/meetup/swarm-sasstools/blob/master/scss/utils/helpers/_i18n.scss
 		const lang = localeCode.substring(0, 2);
 
-		return (
-			<div
-				id="root"
-				className={`column lang_${lang}`}
-				style={{ minHeight: '100vh' }}
-			>
-				<Helmet defaultTitle="Meetup" titleTemplate="%s - Meetup">
-					<link rel="stylesheet" type="text/css" href={webfontCSSHref} />
-					<link rel="stylesheet" type="text/css" href={baseCSSHref} />
-					<meta name="viewport" content="width=device-width, initial-scale=1" />
-					<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-					<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-					<meta name="robots" content="index,follow" />
-					<meta
-						name="verify-v1"
-						content="h5EhuAEkLFlZmMxwpH5wnRaoDEmqYCCEUE+FLcrRNvE="
-					/>
+		/*
 					<link
 						rel="apple-touch-icon"
 						sizes="120x120"
@@ -97,6 +63,29 @@ class PageWrap extends React.Component {
 						href={swarmIcon180x180}
 					/>
 					<link rel="shortcut icon" sizes="196x196" href={swarmIcon196x196} />
+*/
+
+		return (
+			<div
+				id="root"
+				className={`column lang_${lang}`}
+				style={{ minHeight: '100vh' }}
+			>
+				<Helmet defaultTitle="Meetup" titleTemplate="%s - Meetup">
+					{webfontCSSHref &&
+						<link rel="stylesheet" type="text/css" href={webfontCSSHref} />}
+
+					{baseCSSHref &&
+						<link rel="stylesheet" type="text/css" href={baseCSSHref} />}
+
+					<meta name="viewport" content="width=device-width, initial-scale=1" />
+					<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+					<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+					<meta name="robots" content="index,follow" />
+					<meta
+						name="verify-v1"
+						content="h5EhuAEkLFlZmMxwpH5wnRaoDEmqYCCEUE+FLcrRNvE="
+					/>
 
 					<script type="text/javascript" src={polyfillServiceUrl(localeCode)} />
 					<script type="text/javascript">
@@ -104,10 +93,11 @@ class PageWrap extends React.Component {
 					</script>
 				</Helmet>
 
-				<div
-					style={iconSpriteStyle}
-					dangerouslySetInnerHTML={{ __html: iconSprite }}
-				/>
+				{iconSprite &&
+					<div
+						style={{ display: 'none' }}
+						dangerouslySetInnerHTML={{ __html: iconSprite }}
+					/>}
 
 				{this.renderChildren()}
 			</div>
@@ -119,6 +109,9 @@ PageWrap.propTypes = {
 	localeCode: PropTypes.string.isRequired,
 	self: PropTypes.object.isRequired,
 	location: PropTypes.object.isRequired,
+	baseCSSHref: PropTypes.string,
+	webfontCSSHref: PropTypes.string,
+	iconSprite: PropTypes.string,
 };
 
 export default PageWrap;
